requirements.txt
streamlit
requests
pandas
streamlit-agraph
streamlit-authenticator
pickle-mixin
PyPDF2
docx2txt
streamlit_authenticator
google-cloud
grpc-google-cloud-vision-v1
gax-google-cloud-vision-v1
google-cloud-vision
google-cloud-bigquery
Pillow
graphviz
pyvis
python-dotenv
